+++ .github/workflows/workflow.yml
@@ -106,7 +106,7 @@ jobs:
       - name: Build - custommutator
         timeout-minutes: 40
         run: |
-          cd modules/custommutator && make
+          cd custommutator && make
 
       - name: Build - rust bitcoin
         timeout-minutes: 5
@@ -171,6 +171,7 @@ jobs:
         with:
           path: |
             modules/
+            custommutator/
             bitcoinfuzz
             NBitcoin.CppBridge.so
             NBitcoin.CppBridge.dylib
