+++ custommutator/Makefile
@@ -4,7 +4,9 @@ CXXFLAGS += -Wall -Wextra -std=c++20 \
             -fsanitize=fuzzer,address \
             -I . \
             -I ../../include \
-            -I /opt/homebrew/include
+            -I /opt/homebrew/include \
+						-I ./utils \
+						-I ../
 
 MAKE = make
 AR := ar
@@ -13,14 +15,14 @@ AR := ar
 BITCOIN_SRCS := $(shell find . -name "*.cpp" ! -name "bech32.cpp" ! -name "cleanse.cpp" ! -path "./univalue/*" ! -path "./secp256k1/*")
 BITCOIN_OBJS := $(BITCOIN_SRCS:.cpp=.o)
 
-module.a: bech32.o  customcrossover.o $(BITCOIN_OBJS)
-	$(AR) rcs $@ bech32.o customcrossover.o $(BITCOIN_OBJS)
+module.a: utils/bech32.o  mutators/customcrossover.o $(BITCOIN_OBJS)
+	$(AR) rcs $@ utils/bech32.o mutators/customcrossover.o $(BITCOIN_OBJS)
 
-bech32.o: bech32.cpp bech32.h
-	$(CXX) $(CXXFLAGS) -I . -fPIC -c bech32.cpp -o bech32.o
+utils/bech32.o: utils/bech32.cpp utils/bech32.h
+	$(CXX) $(CXXFLAGS) -I . -fPIC -c utils/bech32.cpp -o utils/bech32.o
 
-customcrossover.o: customcrossover.cpp customcrossover.h
-	$(CXX) $(CXXFLAGS) -I . -fPIC -c customcrossover.cpp -o customcrossover.o
+mutators/customcrossover.o: mutators/customcrossover.cpp mutators/customcrossover.h
+	$(CXX) $(CXXFLAGS) -I . -fPIC -c mutators/customcrossover.cpp -o mutators/customcrossover.o
 
 %.o: %.cpp
 	@mkdir -p $(dir $@)
